BLE Host - Send correct error code for unsup. n/i.
Previously, if a peer attempted to enable notifications or indications
for a characteristic that did not support them, the elicited error
response had an error code of 3 (BLE_ATT_ERR_WRITE_NOT_PERMITTED).

Now, the error response has a code of 6 (BLE_ATT_ERR_REQ_NOT_SUPPORTED).

The spec does not appear to indicate the correct error t osend in this
case.  I believe req-not-supported is more correct than
write-not-permitted.